int is replaced to double

+++ additionalTasks/Second.java
@@ -2,11 +2,11 @@ package skillup.additionalTasks;
 
 public class Second {
     public static void main(String args[]) {
-        int a = 1;
-        int b = 2;
-        int c = 5;
-        int sum = a + b + c;
-        int mult = a * b * c;
+        double a = 1;
+        double b = 2;
+        double c = 5;
+        double sum = a + b + c;
+        double mult = a * b * c;
         if (sum > mult) {
             System.out.println("Result of addition is greater then result of multiplication");
         } else if (sum < mult) {
